chore(storybook): modernize Storybook setup
